adding release_tools.py
